backend/requirements.txt
# WebRPA 后端依赖
# 安装命令: pip install -r requirements.txt

# Web框架
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.16.1
python-multipart>=0.0.12
pydantic>=2.10.0
aiofiles>=24.1.0

# 浏览器自动化
playwright>=1.49.0

# 数据处理
# 使用 rtcompat 版本以兼容老旧 CPU（不支持 AVX2 指令集）
# 如果你的 CPU 支持 AVX2，可以使用普通版本以获得更好的性能：polars>=1.17.0
# 如果遇到 "Illegal instruction" 错误，说明需要使用 rtcompat 版本
polars[rtcompat]>=1.17.0
pandas>=3.0.0
pyarrow>=23.0.0
xlsxwriter>=3.2.0
xlrd>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21

# HTML 转 Markdown（Firecrawl 模块）
html2text>=2024.2.26
beautifulsoup4>=4.12.0

# HTTP请求
httpx>=0.27.0

# JSON路径解析（API触发器模块）
jsonpath-ng>=1.6.0

# 正则表达式增强（支持可变长度后向断言）
regex>=2024.0.0

# 验证码识别
# ddddocr 1.0.6 是 Python 3.13 唯一支持的版本
ddddocr>=1.0.6

# 图像识别（点击图像模块）
opencv-python>=4.8.0
Pillow>=10.0.0

# 文档转换（Pandoc模块 - 需要单独安装Pandoc: https://pandoc.org/installing.html）
pypandoc>=1.13

# 图像处理增强（Pillow扩展模块）
pillow-heif>=0.18.0  # HEIF/HEIC格式支持
piexif>=1.1.3  # EXIF元数据处理

# 屏幕捕获（屏幕共享模块）
mss>=9.0.0

# 系统通知（系统消息模块）
plyer>=2.1.0

# Windows API（多显示器截图支持）
pywin32>=306; sys_platform == 'win32'

# 进程管理（浏览器进程清理）
psutil>=5.9.0

# 数据库
pymysql>=1.1.0

# 音频播放（支持几乎所有常见格式：mp3, wav, flac, aac, m4a, ogg, wma, opus等）
# 需要将 ffmpeg.exe 和 ffprobe.exe 放到 backend 目录
pydub>=0.25.0

# 键鼠监听（宏录制器模块）
pynput>=1.7.0

# 键鼠模拟（微信自动化模块）
pyautogui>=0.9.54

# 代理抓包（用于抓取模拟器/手机APP的网络请求）
mitmproxy>=10.0.0

# 人脸识别（人脸触发器模块）
# 注意：face_recognition 依赖 dlib，安装可能需要 CMake 和 Visual Studio Build Tools
# Windows用户安装步骤：
# 1. 安装 CMake: pip install cmake
# 2. 安装 Visual Studio Build Tools (https://visualstudio.microsoft.com/downloads/)
# 3. 安装 dlib: pip install dlib
# 4. 安装 face_recognition: pip install face_recognition
# 或者使用预编译的wheel: https://github.com/z-mahmud22/Dlib_Windows_Python3.x
face_recognition>=1.3.0

# PDF处理
# 注意：pdf2docx 使用 GPL v3 许可证，仅供非商业使用
# 商业使用需要遵守 GPL v3 许可证或联系原作者获取商业授权
pypdf>=4.0.0  # MIT 许可证 - 用于基本PDF操作
reportlab>=4.0.0  # BSD 许可证 - 用于PDF生成
pdf2image>=1.16.0  # MIT 许可证 - PDF转图片（需要安装 poppler）
python-docx>=1.0.0  # MIT 许可证 - Word文档操作
pdfplumber>=0.11.0  # MIT 许可证 - PDF内容提取（文本、图片、表格）
pdf2docx>=0.5.9  # GPL v3 许可证 - 高保真PDF转Word（仅供非商业使用）

# 二维码生成与解码
qrcode>=7.4.0
pyzbar>=0.1.9

# 语音识别（音频转文本）
SpeechRecognition>=3.10.0

# OCR文字识别（点击文本、悬停文本模块）
paddleocr>=2.7.0

# Cron表达式解析（定时触发器模块）
croniter>=2.0.0

# 任务调度（计划任务模块）
APScheduler>=3.10.0

# 手势识别（手势触发器模块）
# MediaPipe 提供高性能的手势识别能力
mediapipe>=0.10.9

# AI爬虫（智能爬虫模块）
scrapegraphai>=1.73.0
nest-asyncio>=1.6.0

# 语音识别（高级）- 音频转文本模块
faster-whisper>=1.2.0

# 测试报告（Allure模块）
# 注意：生成HTML报告需要单独安装 Allure 命令行工具
# 安装命令：npm install -g allure-commandline
# 或访问：https://docs.qameta.io/allure/
# allure-python-commons 仅用于生成测试结果数据
allure-python-commons>=2.13.0

# 桌面应用自动化（Desktop模块）
# uiautomation 提供基于 UI Automation 的桌面应用自动化能力
# 支持Windows桌面应用的UI元素识别和操作
uiautomation>=2.0.0
# pywinauto 提供Windows桌面应用自动化能力
# 支持Win32和UIA两种后端，可以操作几乎所有Windows桌面应用
pywinauto>=0.6.8

# 多渠道通知（Apprise模块）
# Apprise 支持80+种通知服务，包括Discord、Telegram、钉钉、企业微信、飞书等
apprise>=1.9.0

# 数据库驱动（多数据库支持模块）
# Oracle数据库（使用新版oracledb替代cx-Oracle）
oracledb>=3.4.0
# PostgreSQL数据库
psycopg2-binary>=2.9.0
# MongoDB数据库
pymongo>=4.6.0
# SQL Server数据库
pyodbc>=5.0.0
# SQLite数据库（Python内置，无需安装）
# Redis数据库
redis>=5.0.0

# SSH远程操作（SSH模块）
paramiko>=3.4.0

# 飞书API（飞书自动化模块）
# 飞书开放平台SDK
lark-oapi>=1.2.0

# AI图像生成（AI生图模块）
# 支持多种AI服务提供商的统一接口
openai>=1.12.0  # OpenAI DALL-E
# 注意：其他AI服务（如Midjourney、Stable Diffusion）通过API调用，无需额外依赖

# AI视频生成（AI生视频模块）
# 通过API调用各大AI视频服务，无需额外依赖
